Split SPOSetBuilderBase.h from BasisSetBase.h
