+++ src/QMCWaveFunctions/lcao/LCAOrbitalBuilder.h
@@ -19,6 +19,7 @@
 #define QMCPLUSPLUS_SOA_LCAO_ORBITAL_BUILDER_H
 
 #include "QMCWaveFunctions/BasisSetBase.h"
+#include "QMCWaveFunctions/SPOSetBuilderBase.h"
 
 namespace qmcplusplus
 {
